% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Champion.R
\name{lolstat_champions}
\alias{lolstat_champions}
\title{lolstat champions
get information about champion Id and name using DDragon API}
\usage{
lolstat_champions()
}
\value{
data frame with consists of 2 column, champion ID & champion Name in English
}
\description{
lolstat champions
get information about champion Id and name using DDragon API
}
\examples{
champInfo = lolstat_champions()

}
