% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Champion.R
\name{lolstat_champions_id2name}
\alias{lolstat_champions_id2name}
\title{lolstat champions id2name}
\usage{
lolstat_champions_id2name(matches, champInfo)
}
\arguments{
\item{matches}{result of lolstat_matchlists()}

\item{champInfo}{result of lolstat_champions()}
}
\description{
change champion from id to name
}
\examples{
matches = lolstat_champions_id2name(matches, champInfo)

}
