% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{lolstat_game}
\alias{lolstat_game}
\title{lolstat Game}
\usage{
lolstat_game(matchID, region, APIKey)
}
\arguments{
\item{matchID}{identifier of match}

\item{region}{Character, league region}

\item{APIKey}{Character, if set explicitly; not needed if key is set globally. See \code{\link{setAPIKey}}.}
}
\description{
this will return entire match object
}
