% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matchlist.R
\name{lolstat_matchlists}
\alias{lolstat_matchlists}
\title{lolstat Matchlist}
\usage{
lolstat_matchlists(
  accountIds,
  region,
  APIKey,
  champion = NULL,
  queue = NULL,
  endTime = NULL,
  beginTime = NULL,
  endIndex = NULL,
  beginIndex = NULL
)
}
\arguments{
\item{region}{Character, league region}

\item{APIKey}{Character, if set explicitly; not needed if key is set globally. See \code{\link{setAPIKey}}.}

\item{champion}{Numeric value : See \code{\link{lolstat_champions}}.}

\item{queue}{Numeric value 420 : Solo Rank, 430 : Normal, 440 : Team Rank, if NULL : all of these are applied.}

\item{endTime}{Numeric value in epoch mileseconds. refer official API page}

\item{beginTime}{Numeric value in epoch mileseconds. refer official API page}

\item{endIndex}{Numeric value to howmuch of matchlist will be gathered, default is beginIndex + 100}

\item{beginIndex}{Numeric value to howmuch of matchlist will be gathered, default is 0 and should be < 100.}

\item{accountId}{Character, See \code{\link{lolstat_league_summoners}}.}
}
\description{
get all matchlists
}
\details{
for more information, refer official API page, https://developer.riotgames.com/apis#match-v4
}
\examples{
\dontrun{
 # use setAPIKey() to set APIKey
 setAPIKey('myAPIKey', 'sysenv')

 # getAPIKey() to get APIKey
 APIKey = getAPIKey('sysenv')

 # get recent matchlist (up to 100) of first Platinum I summoners matchlist in KR region.
 # P1 = lolstat_league_summoners('I','PLATINUM', 'RANKED_SOLO_5x5', 1, 'KR', APIKey)
 matches = lolstat_matchlists(P1$accountId, 'KR', APIKey )
}

}
